requirements.txt
pymupdf>=1.24.0
Pillow>=10.0.0
ttkbootstrap>=1.10.1
pytesseract>=0.3.10
